refactor(move maskforphone.ts): moved file to /utils and delete directory mask
