+++ src/components/ui-components/Input/config.ts
@@ -4,7 +4,7 @@ import {
   EmailValidate, MaxLengthValidate, NameValidate, PasswordValidate, PhoneValidate,
 } from '@/pages/SignUpPage/options';
 import { capitalizeFirstLetter } from '@/utils/capitalizeFirstLetter';
-import { maskForPhone } from '@/utils/mask/maskForPhone';
+import { maskForPhone } from '@/utils/maskForPhone';
 
 export const NameInputConfig = {
   name: INPUT_FORM_NAMES.NAME,
